+++ src/estimation/struct_estimation/debug_estimation.py
@@ -5,16 +5,13 @@ from set_paths import create_path_dict
 paths_dict = create_path_dict()
 
 import jax
-import yaml
 import pickle as pkl
-import numpy as np
 
 jax.config.update("jax_enable_x64", True)
 
 from set_paths import create_path_dict
 
 path_dict = create_path_dict()
-from estimation.estimate_setup import create_job_offer_params_from_start
 
 # %%
 params = pkl.load(open(path_dict["est_params"], "rb"))
@@ -26,8 +23,6 @@ from model_code.stochastic_processes.policy_states_belief import (
 from model_code.stochastic_processes.policy_states_belief import (
     update_specs_exp_ret_age_trans_mat,
 )
-
-from model_code.specify_model import specify_model
 
 # # Generate model_specs
 # model, params = specify_model(
@@ -51,7 +46,10 @@ solution, model, params = specify_and_solve_model(
 )
 
 
-from estimation.estimate_setup import create_ll_from_paths, load_and_prep_data
+from estimation.struct_estimation.estimate_setup import (
+    create_ll_from_paths,
+    load_and_prep_data,
+)
 
 data_decision, states_dict = load_and_prep_data(
     path_dict, params, model, drop_retirees=True
