+++ src/estimation/struct_estimation/min_class.py
