+++ src/estimation/struct_estimation/server_est.sh
