+++ src/simulation/debug_sim.py
@@ -5,7 +5,6 @@ from set_paths import create_path_dict
 paths_dict = create_path_dict()
 
 import jax
-import yaml
 import pickle as pkl
 import numpy as np
 from model_code.wealth_and_budget.budget_equation import budget_constraint
@@ -15,7 +14,6 @@ jax.config.update("jax_enable_x64", True)
 from set_paths import create_path_dict
 
 path_dict = create_path_dict()
-from estimation.estimate_setup import create_job_offer_params_from_start
 
 # %%
 params = pkl.load(open(path_dict["est_params"], "rb"))
@@ -35,11 +33,6 @@ solution_subj, model, params = specify_and_solve_model(
     policy_state_trans_func=expected_SRA_probs_estimation,
     load_model=True,
     load_solution=True,
-)
-from simulation.policy_state_scenarios.step_function import (
-    create_update_function_for_scale,
-    create_update_function_for_slope,
-    realized_policy_step_function,
 )
 
 # solution_00, model, params = specify_and_solve_model(
